+++ libraries/common/src/main/java/androidx/media3/common/MediaMetadata.java
@@ -80,6 +80,7 @@ public final class MediaMetadata {
     @Nullable private Integer releaseMonth;
     @Nullable private Integer releaseDay;
     @Nullable private CharSequence writer;
+    @Nullable private CharSequence author;
     @Nullable private CharSequence composer;
     @Nullable private CharSequence conductor;
     @Nullable private Integer discNumber;
@@ -123,6 +124,7 @@ public final class MediaMetadata {
       this.releaseMonth = mediaMetadata.releaseMonth;
       this.releaseDay = mediaMetadata.releaseDay;
       this.writer = mediaMetadata.writer;
+      this.author = mediaMetadata.author;
       this.composer = mediaMetadata.composer;
       this.conductor = mediaMetadata.conductor;
       this.discNumber = mediaMetadata.discNumber;
@@ -384,6 +386,14 @@ public final class MediaMetadata {
     @CanIgnoreReturnValue
     public Builder setWriter(@Nullable CharSequence writer) {
       this.writer = writer;
+      return this;
+    }
+
+    /** Sets the author. */
+    @CanIgnoreReturnValue
+    @UnstableApi
+    public Builder setAuthor(@Nullable CharSequence author) {
+      this.author = author;
       return this;
     }
 
@@ -1119,6 +1129,9 @@ public final class MediaMetadata {
   /** Optional writer. */
   @Nullable public final CharSequence writer;
 
+  /** Optional author. */
+  @UnstableApi @Nullable public final CharSequence author;
+
   /** Optional composer. */
   @Nullable public final CharSequence composer;
 
@@ -1203,6 +1216,7 @@ public final class MediaMetadata {
     this.releaseMonth = builder.releaseMonth;
     this.releaseDay = builder.releaseDay;
     this.writer = builder.writer;
+    this.author = builder.author;
     this.composer = builder.composer;
     this.conductor = builder.conductor;
     this.discNumber = builder.discNumber;